Rename the samples to not have an extra dot.
