+++ samples/Samples.AspNetCoreWebApi21/Properties/launchSettings.json
@@ -1,10 +1,10 @@
 ﻿{
   "$schema": "http://json.schemastore.org/launchsettings.json",
   "iisSettings": {
-    "windowsAuthentication": false,
-    "anonymousAuthentication": true,
+    "windowsAuthentication": false, 
+    "anonymousAuthentication": true, 
     "iisExpress": {
-      "applicationUrl": "http://localhost:54599",
+      "applicationUrl": "http://localhost:12372",
       "sslPort": 0
     }
   },
@@ -28,7 +28,7 @@
       "use64Bit": true,
       "nativeDebugging": true
     },
-    "Samples.AspNetCore.WebApi31": {
+    "Samples.AspNetCoreWebApi21": {
       "commandName": "Project",
       "launchBrowser": false,
       "environmentVariables": {
@@ -45,7 +45,7 @@
         "DD_DOTNET_TRACER_HOME": "$(ProjectDir)$(OutputPath)profiler-lib",
         "DD_INTEGRATIONS": "$(ProjectDir)$(OutputPath)profiler-lib\\integrations.json"
       },
-      "applicationUrl": "https://localhost:54598",
+      "applicationUrl": "http://localhost:12373",
       "nativeDebugging": true
     }
   }
