+++ samples/Samples.AspNetCoreWebApi21/Samples.AspNetCoreWebApi21.csproj
