+++ samples/Samples.AspNetCoreWebApi21/appsettings.Development.json
