+++ samples/Samples.AspNetCoreWebApi31/Samples.AspNetCoreWebApi31.csproj
