+++ samples/Samples.AspNetCoreWebApi31/appsettings.Development.json
