+++ samples/Samples.AspNetCoreWebApi31/appsettings.json
